pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "debugpy>=1.8.12",
    "fastapi>=0.115.8",
    "langchain-community>=0.3.17",
    "langchain-openai>=0.3.5",
    "langgraph>=0.2.72",
    "langgraph-checkpoint-postgres>=2.0.14",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.34.0",
]
